refactor create-order page
